caption/requirements.txt
# Required packages for Computer Vision learning
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Optional: Deep Learning frameworks (choose one or both)
# tensorflow>=2.13.0
# torch>=2.0.0
# torchvision>=0.15.0
